+++ src/screens/auth/authHome/AuthHomeScreen.tsx
@@ -5,7 +5,7 @@ import { Image, View } from "react-native";
 import DefaultInput from "@/components/common/input/DefaultInput";
 import { useEffect, useState } from "react";
 import PasswordInput from "@/components/common/input/PasswordInput";
-import Button from "@/components/common/button/Button";
+import Button from "@/components/common/button/defaultButton/DefaultButton";
 import btn_kakaoLogin_default from "@/assets/btn_kakao-login_default.png"
 import btn_naverLogin_default from "@/assets/btn_naver-login_default.png"
 import btn_googleLogin_default from "@/assets/btn_google-login_default.png"
